move files, tweak for play object
