+++ vimeo-networking/src/main/java/com/vimeo/networking/model/playback/embed/EmbedButtons.java
@@ -22,7 +22,7 @@
  * SOFTWARE.
  */
 
-package com.vimeo.networking.model.embed;
+package com.vimeo.networking.model.playback.embed;
 
 import java.io.Serializable;
