+++ vimeo-networking/src/main/java/com/vimeo/networking/model/playback/embed/EmbedHtml.java
@@ -22,44 +22,37 @@
  * SOFTWARE.
  */
 
-package com.vimeo.networking.model.embed;
+package com.vimeo.networking.model.playback.embed;
 
 import org.jetbrains.annotations.Nullable;
 
 import java.io.Serializable;
 
 /**
- * Logo data contained in {@link Embed} objects. These are here to allow you to edit a video's embed settings.
+ * Representation of the html data within a {@link Embed} object. This contains the necessary iframe html for
+ * embedding. This is for getting
  * <p/>
  * Created by zetterstromk on 4/25/16.
  */
-public class EmbedLogos implements Serializable {
+public class EmbedHtml implements Serializable {
 
-    private static final long serialVersionUID = -6946192139468749658L;
-    private boolean vimeo;
+    private static final long serialVersionUID = 3752755790501317766L;
+
+    private int width;
+    private int height;
     @Nullable
-    private EmbedCustomLogos custom;
+    private String html;
 
-    public boolean isVimeo() {
-        return vimeo;
+    public int getWidth() {
+        return width;
     }
 
-    /**
-     * @param vimeo Show or hide the vimeo logo
-     */
-    public void setVimeo(boolean vimeo) {
-        this.vimeo = vimeo;
+    public int getHeight() {
+        return height;
     }
 
     @Nullable
-    public EmbedCustomLogos getCustom() {
-        return custom;
-    }
-
-    /**
-     * @param custom {@link EmbedCustomLogos}
-     */
-    public void setCustom(@Nullable EmbedCustomLogos custom) {
-        this.custom = custom;
+    public String getHtml() {
+        return html;
     }
 }
